+++ QueenCam/QueenCam/Presentation/Views/Common/Settings/Main/SettingsState.swift
@@ -1,0 +1,10 @@
+//
+//  SettingsState.swift
+//  QueenCam
+//
+//  Created by 임영택 on 7/25/26.
+//
+
+struct SettingsState: Equatable {
+  var saveGuidingOverlayImageOn: Bool
+}